dataset/requirements.txt
# Core dependencies
requests>=2.31.0

# Scheduling
APScheduler>=3.10.4

# Configuration
PyYAML>=6.0.1

# Data validation
python-dateutil>=2.8.2